# fix: Improve Big Sur browser detection performance