--- a/packages/launcher/lib/darwin/detection-workaround.js
+++ b/packages/launcher/lib/darwin/detection-workaround.js
@@ -1,0 +1,6 @@
+const { detect } = require('../detect.ts')
+
+detect(undefined, false).then((browsers) => {
+  // eslint-disable-next-line no-console
+  console.log(JSON.stringify(browsers, null, 2))
+})