--- a/packages/launcher/lib/darwin/detection-workaround.js
+++ b/packages/launcher/lib/darwin/detection-workaround.js
@@ -1,4 +1,4 @@
-const { detect } = require('../detect.ts')
+const { detect } = require('../detect')
 
 detect(undefined, false).then((browsers) => {
   // eslint-disable-next-line no-console