--- a/packages/launcher/lib/windows/index.ts
+++ b/packages/launcher/lib/windows/index.ts
@@ -1,4 +1,4 @@
-import fse from 'fs-extra'
+import * as fse from 'fs-extra'
 import os from 'os'
 import { join, normalize, win32 } from 'path'
 import { tap, trim, prop } from 'ramda'