--- a/packages/server/lib/util/find_system_node.js
+++ b/packages/server/lib/util/find_system_node.js
@@ -79,5 +79,6 @@ function findNodePathAndVersion () {
 }
 
 module.exports = {
+  findNodeInFullPath,
   findNodePathAndVersion,
 }