<!-- Thanks for contributing! PLEASE...
- Read our contributing guidelines: https://github.com/cypress-io/cypress/blob/develop/CONTRIBUTING.md 
- Read our Code Review Checklist on coding standards and what needs to be done before a PR can be merged: https://github.com/cypress-io/cypress/blob/develop/CONTRIBUTING.md#Code-Review-Checklist
- Mark this PR as "Draft" if it is not ready for review.
- Make sure you set the correct base branch based on what packages you're changing: https://github.com/cypress-io/cypress/blob/develop/CONTRIBUTING.md#branches
-->

- Closes #17773

### User facing changelog

- Projects will now open much faster on MacOS Big Sur with improved browser detection performance.

### Additional details

The root of the issue is [this bug in Electron](https://github.com/electron/electron/issues/26143), which causes spawning child processes to take 300-1000ms instead of the usual <30ms when using MacOS Big Sur and Electron 12+. We spawn a child process for each binary of each browser, plus potential fallback lookups if any of the standard ones fail. This causes the browser detection on Big Sur to take upwards of 10 seconds.

We considered using the Electron patch in that issue, but it's not a perfect solution (it was reverted in libuv due to failing tests), and would require us to build Electron ourselves, which is a large amount of work and responsibility.

As a workaround, we spawn a single child process (only incurring the performance penalty once instead of N+ times), and use it to call into our browser detection code from the user's Node.js. Since it's outside Electron, it doesn't suffer from the same bug and performs quickly. This reduces the time taken for browser detection from ~10s to ~1s.

### How has the user experience changed?

Users on Big Sur will see projects open much quicker.

### PR Tasks
<!-- These tasks must be completed before a PR is merged.
Delete tasks if they are not applicable. -->

- [x] Have tests been added/updated?
- [x] Has the original issue or this PR been tagged with a release in ZenHub? <!-- (internal team only)-->
- N/A Has a PR for user-facing changes been opened in [`cypress-documentation`](https://github.com/cypress-io/cypress-documentation)? <!-- Link to PR here -->
- N/A Have API changes been updated in the [`type definitions`](https://github.com/cypress-io/cypress/blob/develop/cli/types/cypress.d.ts)?
- N/A Have new configuration options been added to the [`cypress.schema.json`](https://github.com/cypress-io/cypress/blob/develop/cli/schema/cypress.schema.json)?