@bahmutov 2021-08-19
👏 👏 👏 

Sent from my iPhone

> On Aug 19, 2021, at 10:17, cypress-bot[bot] ***@***.***> wrote:
> 
> ﻿
> Thanks for taking the time to open a PR!
> 
> Create a Draft Pull Request if your PR is not ready for review. Mark the PR as Ready for Review when you're ready for a Cypress team member to review the PR.
> Become familiar with the Code Review Checklist for guidelines on coding standards and what needs to be done before a PR can be merged.
> —
> You are receiving this because you are subscribed to this thread.
> Reply to this email directly, view it on GitHub, or unsubscribe.