@flotwig 2021-08-19
BTW, was not able to test this on an actual Big Sur install as part of my review, my VM is stuck on Catalina.