@flotwig 2022-03-03
An upstream fix is coming: https://github.com/electron/electron/issues/26143#issuecomment-1057583242